Modified NAS slightly
I was having trouble with the base64 encoded data, so I replaced it with
something that should definitely work.
I also reduced the size of the auth token so there won't be any possible
overflows. Since it's random data it can easily be adjusted in the
future if for some reason it's too small.