Small fixes
- Sometimes the game sends port 0 as the port to forward to. Obviously
this isn't right, so *try* to match up the server with something, even
if it may or may not be right. Hopefully it's right.
- Tried to fix friends games in some games. Metroid Prime: Hunters seems
to look for the dwc_pid for a friend, and sometimes the friend never
does a statechanged, so assume that any state except 2 should be added
to the server list.