Custom PWT Tournament file.
The PWT Tournament file has 12 slots for downloadable tournaments, only
3 of those were used in the English version. The remaining 9 were filled
with the six Asia-only tournaments as well as three of the debug
tournaments.

PWT Tournament format:
http://projectpokemon.org/forums/showthread.php?25121-PWT-Download-Tournaments

One tournament is 4628 bytes.
Last four bytes for both every individual tournament as well as the
whole file are a CRC16 CCITT checksum of all previous bytes and two null
bytes.

The added tournaments are still in Japanese, if someone wants to provide
me good translations for them feel free.